+++ training/automl/tests/test_configs/payload/forecasting_ui_payload.json
@@ -307,7 +307,7 @@
     ],
     "experimentName": "automl_tests",
     "continueRunOnStepFailure": true,
-    "enforceRerun": false,
+    "enforceRerun": true,
     "computeTarget": "gpu-cluster",
     "displayName": "test_forecasting",
     "tags": {}
